refactor jsonrpc to use index.ts
