+++ packages/jellyfish-jsonrpc/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { JsonRpcClient } from '../src/jsonrpc'
+import { JsonRpcClient } from '../src'
 import { RegTestContainer } from '@defichain/testcontainers'
 import { JellyfishClientError, JellyfishError, JellyfishRPCError } from '@defichain/jellyfish-core'
 import nock from 'nock'
