+++ packages/jellyfish-jsonrpc/package.json
@@ -25,8 +25,8 @@
       "name": "DeFiChain Jellyfish Contributors"
     }
   ],
-  "main": "dist/jsonrpc.js",
-  "types": "dist/jsonrpc.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist"
   ],
